4th round of standard compiler fixes (#2984)
* 4th round of standard compiler fixes

* use of fpcM_ERROR_PROCESS_ID_e in gWolfBustersID

+++ include/d/actor/d_a_npc_kn.h
@@ -179,7 +179,7 @@ public:
 
 public:
     void initialize() {
-        memset(&mFootLPos, 0, (int)&field_0xe3c - (int)&mFootLPos);
+        memset(&mFootLPos, 0, (intptr_t)&field_0xe3c - (intptr_t)&mFootLPos);
 
         mFaceMotionSeqMngr.initialize();
         mMotionSeqMngr.initialize();
